querymind/core/tools/web_search.py
"""Async web search tool stub with Pydantic results and retry/concurrency guards.

Real implementation should use the Tavily async client. This stub is defensive:
- If `tavily` is unavailable it returns an empty list.
- Uses an asyncio.Semaphore to limit concurrent external calls.
- Uses tenacity AsyncRetrying for exponential backoff when available.
"""
from typing import List, Optional
import os
import asyncio
import logging

from pydantic import BaseModel

try:
    import tavily  # type: ignore
except Exception:
    tavily = None

try:
    from tenacity import AsyncRetrying, stop_after_attempt, wait_exponential, retry_if_exception_type
except Exception:
    AsyncRetrying = None  # type: ignore

_CONCURRENCY = int(os.getenv("TAVILY_CONCURRENCY", "5"))
_semaphore = asyncio.Semaphore(_CONCURRENCY)
_LOG = logging.getLogger(__name__)


class SearchResult(BaseModel):
    url: str
    title: str
    content: str
    relevance_score: float = 0.0


async def _fetch_from_tavily(query: str, top_k: int = 5) -> List[SearchResult]:
    if tavily is None:
        raise RuntimeError("tavily client not installed")

    # This code attempts to call a hypothetical async Tavily client.
    # Replace with the real Tavily async API when available.
    client = None
    try:
        client = tavily.AsyncClient(api_key=os.getenv("TAVILY_API_KEY"))
    except Exception:
        # If the exact client class differs, attempt a generic constructor
        try:
            client = tavily.Client(api_key=os.getenv("TAVILY_API_KEY"))
        except Exception:
            raise

    # Best-effort mapping — real response shape depends on the Tavily SDK.
    resp = await client.search(query, limit=top_k)
    results: List[SearchResult] = []
    try:
        items = getattr(resp, "results", resp)
        for it in items:
            url = getattr(it, "url", getattr(it, "link", ""))
            title = getattr(it, "title", "")
            content = getattr(it, "snippet", getattr(it, "content", ""))
            score = float(getattr(it, "score", 0.0) or 0.0)
            results.append(SearchResult(url=url, title=title, content=content, relevance_score=score))
    except Exception:
        _LOG.exception("Failed to parse tavily response; returning empty list")

    return results


async def web_search(query: str, top_k: int = 5) -> List[SearchResult]:
    """Public async web search function.

    - Retries with exponential backoff (tenacity) when available.
    - Concurrency-limited by a semaphore.
    - Returns an empty list when the Tavily client is not installed.
    """
    async with _semaphore:
        if tavily is None or AsyncRetrying is None:
            # Defensive: no external client available in this environment.
            _LOG.debug("Tavily or tenacity not available; returning empty results")
            await asyncio.sleep(0)
            return []

        retryer = AsyncRetrying(
            stop=stop_after_attempt(3),
            wait=wait_exponential(multiplier=1, min=1, max=10),
            retry=retry_if_exception_type(Exception),
            reraise=True,
        )

        async for attempt in retryer:
            with attempt:
                return await _fetch_from_tavily(query, top_k=top_k)


__all__ = ["SearchResult", "web_search"]
from typing import List
import asyncio
from pydantic import BaseModel


class SearchResult(BaseModel):
    url: str
    title: str
    content: str
    relevance_score: float = 0.0


async def web_search(query: str, top_k: int = 5) -> List[SearchResult]:
    """Stub async web_search that returns no results.

    Real implementation should call Tavily (async), apply rate-limiting,
    backoff and return typed `SearchResult` items.
    """
    await asyncio.sleep(0)  # keep function async-friendly
    return []
